add StoreClient to GCS
